Removed old nodeView route and cleaned up prints and imports
