+++ src/main/scala/bifrost/api/http/WalletApiRoute.scala
@@ -11,10 +11,8 @@ import bifrost.transaction.box.{ArbitBox, BifrostBox, PolyBox}
 import bifrost.wallet.BWallet
 import io.circe.Json
 import io.circe.parser.parse
-import io.circe.parser._
 import io.circe.syntax._
 import bifrost.LocalInterface.LocallyGeneratedTransaction
-import bifrost.api.http.{ApiException, SuccessApiResponse}
 import bifrost.settings.Settings
 import bifrost.transaction.box.proposition.{ProofOfKnowledgeProposition, PublicKey25519Proposition}
 import bifrost.transaction.state.PrivateKey25519
